// web/src/app/shared/models/evento.ts
export interface Evento {
  [key: string]: any;
  id?: number;
  nome: string;
  descricao: string;
  local: string;
  data: string;
  dataIniInscricao: string;
  dataFimInscricao: string;
  dataDesconto: string;
  dataValorNormal: string;
  valor1: number;
  valor2: number;
  valorNormal: number;
  pacote2V1: number;
  pacote2V2: number;
  pacote2V3: number;
  pacote3V1: number;
  pacote3V2: number;
  pacote3V3: number;
  pacote4V1: number;
  pacote4V2: number;
  pacote4V3: number;
  pacote1Desc: string;
  pacote2Desc: string;
  pacote3Desc: string;
  pacote4Desc: string;
  pacote1Ativo: number;
  pacote2Ativo: number;
  pacote3Ativo: number;
  pacote4Ativo: number;
  categoria: string;
  boletoInf1: string;
  boletoInf2: string;
  boletoInf3: string;
  boletoInstrucao1: string;
  boletoInstrucao2: string;
  boletoInstrucao3: string;
  obsTela: string;
  txtEmailCadastro: string;
  txtEmailBaixa: string;
  ativaInscricao: number;
  ativaEvento: number;
  ativaAlteracaoInscricao: number;
  eventoTipo: number;
  pacote1V1Pseg: string;
  pacote1V2Pseg: string;
  pacote1V3Pseg: string;
  pacote2V1Pseg: string;
  pacote2V2Pseg: string;
  pacote2V3Pseg: string;
  pacote3V1Pseg: string;
  pacote3V2Pseg: string;
  pacote3V3Pseg: string;
  pacote4V1Pseg: string;
  pacote4V2Pseg: string;
  pacote4V3Pseg: string;
  imagem?: string;
  notifications?: [];
  isValid?: number;
}